# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/analytics/median": {
            "get": {
                "description": "Returns the median(p50) for period",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get median",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start of range (from=2026-01-10T12:15:31+03:00)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End of range   (to=2026-03-22T19:07:13+03:00)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "If you need a specific product",
                        "name": "product-id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "If you need a specific category",
                        "name": "category-id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "If you need a specific payment method",
                        "name": "payment-method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Grouping by different fields (group-by=day/month/week/category/product)",
                        "name": "group-by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting by field (sort-by=group/value)",
                        "name": "sort-by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting order (sort-order=asc/desc)",
                        "name": "sort-order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.AnalyticsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/v1/analytics/p90": {
            "get": {
                "description": "Returns the p90 for period",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get percentile 90",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start of range (from=2026-01-10T12:15:31+03:00)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End of range   (to=2026-03-22T19:07:13+03:00)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "If you need a specific product",
                        "name": "product-id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "If you need a specific category",
                        "name": "category-id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "If you need a specific payment method",
                        "name": "payment-method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Grouping by different fields (group-by=day/month/week/category/product)",
                        "name": "group-by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting by field (sort-by=group/value)",
                        "name": "sort-by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting order (sort-order=asc/desc)",
                        "name": "sort-order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.AnalyticsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/v1/analytics/revenue-avg": {
            "get": {
                "description": "Returns the average revenue for period",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get revenue average",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start of range (from=2026-01-10T12:15:31+03:00)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End of range   (to=2026-03-22T19:07:13+03:00)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "If you need a specific product",
                        "name": "product-id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "If you need a specific category",
                        "name": "category-id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "If you need a specific payment method",
                        "name": "payment-method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Grouping by different fields (group-by=day/month/week/category/product)",
                        "name": "group-by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting by field (sort-by=group/value)",
                        "name": "sort-by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting order (sort-order=asc/desc)",
                        "name": "sort-order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.AnalyticsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/v1/analytics/revenue-sum": {
            "get": {
                "description": "Returns the revenue sum for period",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get revenue sum",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start of range (from=2026-01-10T12:15:31+03:00)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End of range   (to=2026-03-22T19:07:13+03:00)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "If you need a specific product",
                        "name": "product-id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "If you need a specific category",
                        "name": "category-id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "If you need a specific payment method",
                        "name": "payment-method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Grouping by different fields (group-by=day/month/week/category/product)",
                        "name": "group-by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting by field (sort-by=group/value)",
                        "name": "sort-by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting order (sort-order=asc/desc)",
                        "name": "sort-order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.AnalyticsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/v1/analytics/sales-count": {
            "get": {
                "description": "Returns the number of sales for period",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get sales count",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start of range (from=2026-01-10T12:15:31+03:00)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End of range   (to=2026-03-22T19:07:13+03:00)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "If you need a specific product",
                        "name": "product-id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "If you need a specific category",
                        "name": "category-id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "If you need a specific payment method",
                        "name": "payment-method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Grouping by different fields (group-by=day/month/week/category/product)",
                        "name": "group-by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting by field (sort-by=group/value)",
                        "name": "sort-by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting order (sort-order=asc/desc)",
                        "name": "sort-order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.AnalyticsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.AnalyticsResponse": {
            "type": "object",
            "properties": {
                "group_label": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}